refactor: add contexts folder
